Move charts into components/charts folder
