+++ components/charts/ExerciseProgressionChart.tsx
